aisetu_flask/requirements.txt
Flask==3.1.3
Flask-Cors==5.0.0
pymongo[srv]==4.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
requests==2.32.5
